+++ src/main/java/kr/anabada/anabadaserver/domain/save/service/KnowTogetherService.java
@@ -1,8 +1,8 @@
 package kr.anabada.anabadaserver.domain.save.service;
 
 import kr.anabada.anabadaserver.common.service.ImageService;
-import kr.anabada.anabadaserver.domain.save.dto.SaveSearchRequestDto;
 import kr.anabada.anabadaserver.domain.save.dto.request.KnowTogetherRequest;
+import kr.anabada.anabadaserver.domain.save.dto.request.SaveSearchRequestDto;
 import kr.anabada.anabadaserver.domain.save.dto.response.KnowTogetherResponse;
 import kr.anabada.anabadaserver.domain.save.entity.KnowTogether;
 import kr.anabada.anabadaserver.domain.save.entity.Save;
